Issue(dev) - Fix issue with the localstack init script trying to create the DynamoDB tables twice.
Bug description:
When restarting the localstack container, the init script would error due to the fact that the cdk app would bootstrap the application every time.
To fix this, I added a condition to the init script to only run the bootstrap command when the CDKToolkit stack has not been added.
Unfortunately cdk stack are not persisted by the community edition of localstack, so I have had to remove the resistance entirely.